feat: split app into ES modules
