+++ index.html
@@ -64,6 +64,6 @@
   <div id="toastNotification" class="toast" aria-live="assertive" aria-atomic="true">Settings Saved!</div>
   <canvas id="confettiCanvas" style="position:fixed;top:0;left:0;width:100vw;height:100vh;pointer-events:none;z-index:9999;display:none;"></canvas>
    <div id="confettiMsg" role="status" aria-live="polite"></div>
-  <script src="script.js"></script>
+  <script type="module" src="main.js"></script>
 </body>
 </html>
